Added date moderation endpoints

+++ src/main/java/com/example/demo/repository/CalendarYearRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.CalendarYear;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalendarYearRepository extends JpaRepository<CalendarYear, Short> {
+public interface CalendarYearRepository extends JpaRepository<CalendarYear, Integer> {
 }
